Fix: castear url_temporal_minutos a int

+++ config/almacenamiento.php
@@ -3,7 +3,7 @@
 return [
 
     // Minutos de validez de las URLs firmadas para ver/descargar evidencias privadas
-    'url_temporal_minutos' => env('B2_URL_TEMPORAL_MINUTOS', 60),
+    'url_temporal_minutos' => (int) env('B2_URL_TEMPORAL_MINUTOS', 60),
 
     // Carpeta base dentro del bucket para las evidencias de Equipos
     'carpeta_evidencias' => 'evidencias',
